Update EasyMorse.cpp to conform to morse codes.
I am using the fabulous EasyMorse lib for a very small project to raise interest in morse stuff to play around and earn a certificate if you can manage to morse you own name. I tried it out with the morse sheet from wikipedia but some letters did noch match. There might be reasons for not using the standard morse codes but maybe it was just too much dit-dah-dit-dit ... while setting it up :-)